@@ -5,7 +5,9 @@
     "start": "webpack-dev-server --hot --progress"
   },
   "dependencies": {
+    "apexcharts": "^4.1.0",
     "react": "^18.1.0",
+    "react-apexcharts": "^1.6.0",
     "react-dom": "^18.1.0"
   },
   "devDependencies": {
@@ -16,33 +18,30 @@
     "@babel/preset-react": "^7.22.5",
     "@babel/preset-typescript": "^7.22.5",
     "@babel/runtime": "^7.22.6",
-    "@types/jest": "^29.5.2",
-    "babel-jest": "^29.6.0",
+    "@testing-library/jest-dom": "6.6.3",
+    "@testing-library/react": "16.0.1",
+    "@types/jest": "29.5.14",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "autoprefixer": "^10.4.20",
+    "babel-jest": "29.7.0",
     "babel-loader": "^9.1.2",
+    "css-loader": "^7.1.2",
     "html-webpack-plugin": "^5.5.3",
-    "jest": "^29.6.0",
+    "jest": "29.7.0",
+    "jest-canvas-mock": "2.5.2",
+    "jest-environment-jsdom": "29.7.0",
     "jest-transform-stub": "^2.0.0",
-    "mini-css-extract-plugin": "^2.7.6",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.1.6",
+    "mini-css-extract-plugin": "^2.9.2",
+    "postcss": "^8.4.49",
+    "postcss-loader": "^8.1.1",
+    "style-loader": "^4.0.0",
+    "tailwindcss": "^3.4.15",
+    "ts-jest": "29.2.5",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.7.2",
     "webpack": "^5.88.1",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1"
-  },
-  "jest": {
-    "testRegex": "\\.(test|spec)\\.(j|t)sx?$",
-    "transform": {
-      "^.+\\.jsx?$": "babel-jest",
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "moduleNameMapper": {
-      "^.+.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "jest-transform-stub"
-    },
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx"
-    ]
   }
 }
