@@ -1,7 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';  // Asegúrate de importar desde 'react-dom/client'
 import { App } from './app';
+import './index.css';
 
-document.addEventListener('DOMContentLoaded', () => {
-    ReactDOM.render((<App />), document.getElementById('reactMountPoint'));
-});
+const root = ReactDOM.createRoot(document.getElementById('reactMountPoint')!); 
+root.render(<App />);
