@@ -1,7 +1,16 @@
 import React from 'react';
+import EnergyChart from './components/EnergyChart';
 
-const App = () => <h1>UK Energy Mix</h1>;
-
+function App ()  {
+    return (
+    <div className="flex h-screen">
+        <main className="flex-1 p-6 bg-gray-100 dark:bg-gray-800 dark:text-white overflow-y-auto">
+        <EnergyChart />
+        </main>
+    </div>
+      )
+}
 export {
     App
 }
+
